feat: adjust code to use PBJ generated code
