+++ hapi/src/main/java/module-info.java
@@ -51,17 +51,9 @@ module com.hedera.node.hapi {
     exports com.hedera.hapi.node.state.throttles;
     exports com.hedera.hapi.node.state.congestion;
     exports com.hedera.hapi.platform.event;
-    exports com.hedera.services.stream.proto;
-    exports com.hederahashgraph.api.proto.java;
-    exports com.hederahashgraph.service.proto.java;
     exports com.hedera.hapi.util;
 
-    requires transitive com.google.common;
-    requires transitive com.google.protobuf;
     requires transitive com.hedera.pbj.runtime;
-    requires transitive io.grpc.stub;
-    requires transitive io.grpc;
-    requires io.grpc.protobuf;
     requires org.antlr.antlr4.runtime;
     requires static com.github.spotbugs.annotations;
     requires static java.annotation;
