+++ hedera-node/hapi-utils/build.gradle.kts
@@ -17,6 +17,7 @@
 plugins {
     id("com.hedera.gradle.services")
     id("com.hedera.gradle.services-publish")
+    id("com.hedera.gradle.java-test-fixtures")
 }
 
 description = "Hedera Services API Utilities"
