+++ hedera-node/hapi-utils/src/testFixtures/java/module-info.java
@@ -1,0 +1,7 @@
+module com.hedera.node.app.hapi.utils.test.fixtures {
+    requires com.hedera.node.app.hapi.utils;
+    requires transitive com.hedera.node.hapi;
+    requires transitive com.hedera.node.hapi.test.fixtures;
+    requires transitive com.google.protobuf;
+    requires static com.github.spotbugs.annotations;
+}
